+++ tests/unit/interface/tango/test_track_cmd.py
@@ -181,7 +181,7 @@ def test_set_track_cmd_succeeds_when_dish_mode_is_operate(
     # Force dishManager dishMode to go to OPERATE
     ds_cm._update_component_state(operatingmode=DSOperatingMode.POINT)
     spf_cm._update_component_state(operatingmode=SPFOperatingMode.OPERATE)
-    spfrx_cm._update_component_state(operatingmode=SPFRxOperatingMode.DATA_CAPTURE)
+    spfrx_cm._update_component_state(operatingmode=SPFRxOperatingMode.OPERATE)
     main_event_store.wait_for_value(DishMode.OPERATE)
     ds_cm._update_component_state(pointingstate=PointingState.READY)
     main_event_store.wait_for_value(PointingState.READY)